Update README.md
Project Overview:

The aim of this project is to design an online platform that connects charity organizations with donors and volunteers, facilitating communication between benefactors and those in need. This platform addresses not only monetary donations but also non-monetary needs, such as specialized skills and volunteer work. For instance, some needs require expertise, like caregiving for the elderly or children, or offering educational support through voluntary teaching in underserved areas.

How it Works:

Charity organizations can post their needs as projects on the platform, specifying the requirements for each project. Donors or volunteers can browse these projects and choose the ones that align with their skills and availability. Each project comes with specific details, allowing volunteers to search for opportunities that suit them. Once a volunteer expresses interest, the charity can review their application and either accept or decline it based on the criteria.

Both charities and volunteers can register on the platform. Volunteers can apply to fulfill needs, and organizations can approve or reject these applications. Once a need is met, the organization can mark the project as completed.